DUN-31:Migrate properties to Editor store
creation util schema

Update actions

clean code

CR
